Use Andy's suggestion and extend prmt_fine to handle coarse PRMT as well.
Before this commit, dependencies because of implicit branching in the control
flow weren't handled correctly. These are branches based on match hit/miss or
branches based on what action was executed in a table, where there is no if
statement in the P4 source.

See page 58, section 12 of the P4 spec for examples of such branches:
http://p4.org/wp-content/uploads/2016/11/p4-spec-latest.pdf